Update project billing page to direct to new billing portal
